# public/openApiEspecification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vacation Plan API",
    "description": "Uma simples API para criar eventos e adicionar participantes.",
    "contact": {
      "name": "Fernando Braga Alves",
      "email": "f.braga.alves1@gmail.com"
    },
    "license": {
      "name": "AGPLv3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://localhost/{basePath}",
      "description": "Servidor local para desenvolvimento",
      "variables": {
        "basePath": {
          "default": "api"
        }
      }
    }
  ],
...
  "paths": {
...
    "/holiday-plans": {
      "get": {
        "description": "Rota para listar todos os feriados",
        "tags": [
          "Vacation Plan"
        ],
        "security": [
          {
            "auth": []
          }
        ],
        "operationId": "listHolidayPlans",
        "responses": {
          "200": {
            "description": "Feriados listados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "holiday-plans": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Holiday"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Rota para criar um novo feriado",
        "tags": [
          "Vacation Plan"
        ],
        "operationId": "createHolidayPlan",
        "security": [
          {
            "auth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "date",
                  "location"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "location": {
                    "type": "string"
                  },
                  "participants": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feriado criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Holiday"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "Holiday": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "location": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      }
    },
...
    "securitySchemes": {
      "auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header"
      }
    }
  }
}